Save artifacts under full path rather than relative to the repo root in the artifacts-out folder
Previously, given a a repo like
- repoRoot
  |- subApp1
  |- subApp2

If working on subApp1, and saving the file `fileToSave` it would be stored in the root of the artifactOutFolder, i.e `artifacts-out/fileToSave`. This made it impossible to save artifacts from parent directories.

With the change the file will be saved under `artifacts-out/subApp1/fileToSave`